roster page refactored
