+++ src/app/components/Professor/assignments/assignments.component.scss
@@ -10,27 +10,27 @@
 .assignment-wrapper {
     display: flex;
   }
-  
+
   .assignment-table {
-    width: 100%;
+    width: 95%;
     border-collapse: collapse;
   }
-  
+
   .assignment-table tr:nth-child(even){background-color: #f2f2f2;}
   .assignment-table tr:hover {background-color: #ddd;}
-  
+
   .assignment-table, .assignment-table th {
     border: 1px solid #ddd;
     padding: 8px;
   }
-  
+
   .assignment-table th {
     padding-top: 12px;
     padding-bottom: 12px;
     text-align: left;
     color: white;
   }
-  
+
   .assignment-table td {
     padding: 5px;
   }
