+++ src/app/components/Professor/profile/profile.component.scss
@@ -74,7 +74,7 @@
 }
 
 .roster-table {
-  width: 100%;
+  width: 95%;
   border-collapse: collapse;
 }
 
@@ -103,7 +103,7 @@
 }
 
 .blue-button {
-  margin-bottom: 8px;
+  margin: 8px;
 }
 
 .student-input {
@@ -126,6 +126,6 @@
 
 .add-button {
   position: fixed;
-  bottom: 20px; 
+  bottom: 20px;
   left: 220px;
 }
