+++ src/app/components/Professor/roster/roster.component.scss
@@ -1,0 +1,49 @@
+.roster-wrapper {
+  display: flex;
+}
+
+.roster-table {
+  width: 95%;
+  border-collapse: collapse;
+}
+
+.roster-table tr:nth-child(even){background-color: #f2f2f2;}
+.roster-table tr:hover {background-color: #ddd;}
+
+.roster-table, .roster-table th {
+  border: 1px solid #ddd;
+  padding: 8px;
+}
+
+.roster-table th {
+  padding-top: 12px;
+  padding-bottom: 12px;
+  text-align: left;
+  color: white;
+}
+
+.roster-table td {
+  padding: 5px;
+}
+
+.student-data {
+  justify-content: space-between;
+  display: flex;
+}
+
+.blue-button {
+  margin: 8px;
+}
+
+.student-input {
+  width: 30vw;
+  height: 30vh;
+  padding-top: 10px;
+}
+
+.back-button {
+  width: 4.5em;
+  text-align: center;
+  height: 2.5em;
+  margin-top: 5px;
+}
