+++ src/app/components/Professor/roster/roster.component.spec.ts
@@ -1,0 +1,21 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { RosterComponent } from './roster.component';
+
+describe('RosterComponent', () => {
+  let component: RosterComponent;
+  let fixture: ComponentFixture<RosterComponent>;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      declarations: [RosterComponent]
+    });
+    fixture = TestBed.createComponent(RosterComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
